Resolved circural dependencies issue (#153)
