+++ src/app/frontend/logs/logs_module.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import stateConfig from './logs_state';
+import stateConfig from './logs_stateconfig';
 import {LogColorInversionService} from './logs_service';
 
 /**
