+++ src/app/frontend/logs/logs_state.js
@@ -12,10 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {LogsController} from './logs_controller';
-import LogsToolbarController from './logstoolbar/logstoolbar_controller';
-import {toolbarViewName} from '../chrome/chrome_state';
-
 /** Name of the state. Can be used in, e.g., $state.go method. */
 export const stateName = 'logs';
 
@@ -45,60 +41,4 @@ export class StateParams {
     /** @export {string} Name of this pod container. */
     this.container = container;
   }
-}
-
-/**
- * Configures states for the logs view.
- *
- * @param {!ui.router.$stateProvider} $stateProvider
- * @ngInject
- */
-export default function stateConfig($stateProvider) {
-  let views = {
-    '': {
-      templateUrl: 'logs/logs.html',
-      controller: LogsController,
-      controllerAs: 'ctrl',
-    },
-    [toolbarViewName]: {
-      templateUrl: 'logs/logstoolbar/logstoolbar.html',
-      controller: LogsToolbarController,
-      controllerAs: 'ctrl',
-    },
-  };
-
-  $stateProvider.state(stateName, {
-    url: '/logs/:namespace/:replicaSet/:podId/:container',
-    resolve: {
-      'replicaSetPods': resolveReplicaSetPods,
-      'podLogs': resolvePodLogs,
-    },
-    views: views,
-  });
-}
-
-/**
- * @param {!StateParams} $stateParams
- * @param {!angular.$resource} $resource
- * @return {!angular.$q.Promise}
- * @ngInject
- */
-function resolveReplicaSetPods($stateParams, $resource) {
-  /** @type {!angular.Resource<!backendApi.ReplicaSetPods>} */
-  let resource = $resource('/api/replicasets/pods/:namespace/:replicaSet', $stateParams);
-
-  return resource.get().$promise;
-}
-
-/**
- * @param {!StateParams} $stateParams
- * @param {!angular.$resource} $resource
- * @return {!angular.$q.Promise}
- * @ngInject
- */
-function resolvePodLogs($stateParams, $resource) {
-  /** @type {!angular.Resource<!backendApi.Logs>} */
-  let resource = $resource('/api/logs/:namespace/:podId/:container', $stateParams);
-
-  return resource.get().$promise;
 }
